change proximity algorithm to not compute norm
What we really care bout is not the norm of the trajectory. but the distance from the attractors. and this distance we have already computed. with this change, we not only make the algorithm faster, but also more precise (i.e., what if the typical magnitude of the dynamical system is already 1e6 because the user uses bad normalizations?) @awage