Simplify Lorenz84
So this is a hard system. The continuation works, our recurrences algorithm as always is hard for it because of the attractors being so close and interlaced. Close to the global bifurcation, I see our algorithm is identifying 4 attractors instead of 3, with one being on top of another one. I guess refining further the accuracy would resolve that, but I don't care at the moment! :D