Finalize recurrences seeded continuation (#274)
* correct changelog

* add some code for lorenz84

* correct way to plot attractors

* correct description of what is going on

* current Lorenz84 code never terminates

* add safety counter; IT DOESNT SOLVE TE PROBLEM

* make safety counter ERROR instead.

Things work now. But something is wrong, we are loosing a frcation. I think its because of the NON UNIQUE key assignment into the attractors.

* improve docs of `Sparse`, add unsafe keyword condition

* Fix error with unsafe keyword.
Add Coupled kuramoto example.

* fix everything todo bien

* remove kuramoto (for now!)

* [wip, doesn't work] gotta write tests

* yeeeeeeeeeees fully debug the INSANE dict swapping!

* BUGFIX: show_progress not propagated in mapper

* wip minor alteration

* Simplify Lorenz84

So this is a hard system. The continuation works, our recurrences algorithm as always is hard for it because of the attractors being so close and interlaced. Close to the global bifurcation, I see our algorithm is identifying 4 attractors instead of 3, with one being on top of another one. I guess refining further the accuracy would resolve that, but I don't care at the moment! :D

* port sparse code here

* typos in comments

* fix tests (...?)

* fix all tests!

* reaaaally fix tests

Co-authored-by: awage <>